changed customer name

+++ BinaryFunctions.c
@@ -33,7 +33,7 @@ int saveInventoryToBinaryFile(const Inventory* pInventory, const char* filename)
 
 int writeBeerToBFile(FILE* pFile, const Beer* pBeer)
 {
-    int len = strlen(pBeer->brand) + 1;
+    int len = (int)strlen(pBeer->brand) + 1;
     if (fwrite(&len, sizeof(int), 1, pFile) != 1) return 0;
     if (fwrite(pBeer->brand, sizeof(char), len, pFile) != len) return 0;
     if (fwrite(&pBeer->itemSerial, sizeof(int), 1, pFile) != 1) return 0;
@@ -57,7 +57,7 @@ int writeBeerArrToBFile(FILE* pFile, const Beer* pBeerArr, const int count)
 
 int writeWhiskeyToBFile(FILE* pFile, const Whiskey* pWhiskey)
 {
-    int len = strlen(pWhiskey->brand) + 1;
+    int len = (int)strlen(pWhiskey->brand) + 1;
     if (fwrite(&len, sizeof(int), 1, pFile) != 1) return 0;
     if (fwrite(pWhiskey->brand, sizeof(char), len, pFile) != len) return 0;
     if (fwrite(&pWhiskey->itemSerial, sizeof(int), 1, pFile) != 1) return 0;
@@ -386,7 +386,7 @@ void saveReservationToBinaryFile(const Reservation* reservation, FILE* file)
 {
     fwrite(&reservation->ReservationCode, sizeof(int), 1, file);
 
-    int nameLength = strlen(reservation->customer->name) + 1;
+    int nameLength =(int) strlen(reservation->customer->name) + 1;
     fwrite(&nameLength, sizeof(int), 1, file);
     fwrite(reservation->customer->name, sizeof(char), nameLength, file);
 
@@ -551,7 +551,7 @@ int saveReservationsArrayToBinaryFile(const Sales* pSales, const char* filename)
 int	 writeStringToCompressFile(const char* str, FILE* fp, const char* msg)
 {
 
-    if (!writeCharsToFile(str, strlen(str), fp, msg))
+    if (!writeCharsToFile(str, (int)strlen(str), fp, msg))
         return 0;
